% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapefiles.R
\docType{data}
\name{cc_sp_appeal}
\alias{cc_sp_appeal}
\title{Shapefile with appeal courts coverage}
\format{An object of class SpatialPolygonsDataFrame with 11 polygons for
appeal coverages; projection: "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"}
\source{
Shapefiles from \url{http://www.gis-support.pl/downloads} and own
computation
}
\usage{
cc_sp_appeal
}
\description{
A shapefile with polygons corresponding to coverage of 11 appeal courts.
Ready to plot.
}
\keyword{datasets}

