% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/saosTools.R
\docType{package}
\name{saosTools-package}
\alias{saosTools-package}
\title{Utilities for saos package}
\description{
Utilities for working with \href{saos}{https://github.com/bartekch/saos}
package: shapefiles with coverage of appeal and regional courts, dataset with
precises courts' coordinates, functions for common tasks.
}
\author{
Bartosz Chrol \email{bartek.chrol@gmail.com}
}
\keyword{package}

