% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update_all.R
\name{update_all}
\alias{update_all}
\title{Updating data}
\usage{
update_all(time, update_function)
}
\arguments{
\item{time}{time in POSIXct format passed to 'sinceModificationDate' argument}

\item{update_function}{function which performs actual update, must be provided by
a user}
}
\description{
Update your database (in any format) with the latest judgments loaded into
repository. Basically it only downloads judgments since provided time and
runs updating fucntion, which you have to provide by yourself. See
vignette for an example framework. Note that the time of judgment isn't
necessary identical to the time of loading (or modification) in repository,
therefore this function, makes use of 'sinceModificationDate' argument.
}

